requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2     
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2   
blinker==1.6.2
cachetools==5.3.1        
certifi==2023.5.7        
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.8
decorator==5.1.1
executing==1.2.0
faiss-cpu==1.7.4
Flask==2.3.2
Flask-Cors==4.0.0
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-core==2.3.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
idna==3.4
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
langchain~=0.0.219
langchainplus-sdk==0.0.17
llama-index==0.6.37
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==4.23.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.15.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2